notification with redux finished
